requirements.txt
numpy
flask
xgboost
pandas
sklearn
tqdm
torch=1.6.0+cu101
matplotlib
scikit-learn
torch_geometric
dgl-cu101

